requirements.txt
streamlit>=1.45.0
google-genai>=1.0.0
Pillow>=10.0.0
